requirements.txt
# Core AWS SDK
boto3==1.34.162
botocore==1.34.162

# Data Processing
pandas==2.2.3
numpy==1.26.4
pyarrow==17.0.0  # For efficient data serialization

# Web Framework & API
fastapi==0.115.4
uvicorn[standard]==0.32.0
websockets==13.1
python-multipart==0.0.12

# Data Validation
pydantic==2.9.2
pydantic-settings==2.6.0

# AWS Lambda Specific
aws-lambda-powertools==3.2.0  # Logging, tracing, metrics
mangum==0.19.0  # ASGI adapter for Lambda

# Streaming & Async
aioboto3==13.2.0
aiohttp==3.10.10

# Data Serialization
orjson==3.10.10  # Fast JSON
msgpack==1.1.0

# Monitoring & Logging
structlog==24.4.0
python-json-logger==3.1.0

# Configuration
python-dotenv==1.0.1
pyyaml==6.0.2

# Utilities
python-dateutil==2.9.0
pytz==2024.2
tenacity==9.0.0  # Retry logic
backoff==2.2.1  # Backoff strategies

# StatsBomb Data Handling
statsbombpy==1.11.0  # StatsBomb API wrapper
requests==2.32.3

# Mathematical & Statistical
scipy==1.14.1
scikit-learn==1.5.2

# AWS Infrastructure
aws-cdk-lib==2.160.0
constructs==10.3.0

# Performance & Profiling
memory-profiler==0.61.0
py-spy==0.3.14
